+++ src/class21/Code02_CoinsWayNoLimit.java
@@ -1,55 +1,21 @@
-package class20;
+package class21;
 
-import java.util.HashMap;
-import java.util.Map.Entry;
-
-public class Code03_CoinsWaySameValueSamePapper {
-
-	public static class Info {
-		public int[] coins;
-		public int[] zhangs;
-
-		public Info(int[] c, int[] z) {
-			coins = c;
-			zhangs = z;
-		}
-	}
-
-	public static Info getInfo(int[] arr) {
-		HashMap<Integer, Integer> counts = new HashMap<>();
-		for (int value : arr) {
-			if (!counts.containsKey(value)) {
-				counts.put(value, 1);
-			} else {
-				counts.put(value, counts.get(value) + 1);
-			}
-		}
-		int N = counts.size();
-		int[] coins = new int[N];
-		int[] zhangs = new int[N];
-		int index = 0;
-		for (Entry<Integer, Integer> entry : counts.entrySet()) {
-			coins[index] = entry.getKey();
-			zhangs[index++] = entry.getValue();
-		}
-		return new Info(coins, zhangs);
-	}
+public class Code02_CoinsWayNoLimit {
 
 	public static int coinsWay(int[] arr, int aim) {
 		if (arr == null || arr.length == 0 || aim < 0) {
 			return 0;
 		}
-		Info info = getInfo(arr);
-		return process(info.coins, info.zhangs, 0, aim);
+		return process(arr, 0, aim);
 	}
 
-	public static int process(int[] coins, int[] zhangs, int index, int rest) {
-		if (index == coins.length) {
+	public static int process(int[] arr, int index, int rest) {
+		if (index == arr.length) {
 			return rest == 0 ? 1 : 0;
 		}
 		int ways = 0;
-		for (int zhang = 0; zhang * coins[index] <= rest && zhang <= zhangs[index]; zhang++) {
-			ways += process(coins, zhangs, index + 1, rest - (zhang * coins[index]));
+		for (int zhang = 0; zhang * arr[index] <= rest; zhang++) {
+			ways += process(arr, index + 1, rest - (zhang * arr[index]));
 		}
 		return ways;
 	}
@@ -58,17 +24,14 @@ public class Code03_CoinsWaySameValueSamePapper {
 		if (arr == null || arr.length == 0 || aim < 0) {
 			return 0;
 		}
-		Info info = getInfo(arr);
-		int[] coins = info.coins;
-		int[] zhangs = info.zhangs;
-		int N = coins.length;
+		int N = arr.length;
 		int[][] dp = new int[N + 1][aim + 1];
 		dp[N][0] = 1;
 		for (int index = N - 1; index >= 0; index--) {
 			for (int rest = 0; rest <= aim; rest++) {
 				int ways = 0;
-				for (int zhang = 0; zhang * coins[index] <= rest && zhang <= zhangs[index]; zhang++) {
-					ways += dp[index + 1][rest - (zhang * coins[index])];
+				for (int zhang = 0; zhang * arr[index] <= rest; zhang++) {
+					ways += dp[index + 1][rest - (zhang * arr[index])];
 				}
 				dp[index][rest] = ways;
 			}
@@ -80,20 +43,14 @@ public class Code03_CoinsWaySameValueSamePapper {
 		if (arr == null || arr.length == 0 || aim < 0) {
 			return 0;
 		}
-		Info info = getInfo(arr);
-		int[] coins = info.coins;
-		int[] zhangs = info.zhangs;
-		int N = coins.length;
+		int N = arr.length;
 		int[][] dp = new int[N + 1][aim + 1];
 		dp[N][0] = 1;
 		for (int index = N - 1; index >= 0; index--) {
 			for (int rest = 0; rest <= aim; rest++) {
 				dp[index][rest] = dp[index + 1][rest];
-				if (rest - coins[index] >= 0) {
-					dp[index][rest] += dp[index][rest - coins[index]];
-				}
-				if (rest - coins[index] * (zhangs[index] + 1) >= 0) {
-					dp[index][rest] -= dp[index + 1][rest - coins[index] * (zhangs[index] + 1)];
+				if (rest - arr[index] >= 0) {
+					dp[index][rest] += dp[index][rest - arr[index]];
 				}
 			}
 		}
@@ -104,8 +61,12 @@ public class Code03_CoinsWaySameValueSamePapper {
 	public static int[] randomArray(int maxLen, int maxValue) {
 		int N = (int) (Math.random() * maxLen);
 		int[] arr = new int[N];
+		boolean[] has = new boolean[maxValue + 1];
 		for (int i = 0; i < N; i++) {
-			arr[i] = (int) (Math.random() * maxValue) + 1;
+			do {
+				arr[i] = (int) (Math.random() * maxValue) + 1;
+			} while (has[arr[i]]);
+			has[arr[i]] = true;
 		}
 		return arr;
 	}
@@ -121,7 +82,7 @@ public class Code03_CoinsWaySameValueSamePapper {
 	// 为了测试
 	public static void main(String[] args) {
 		int maxLen = 10;
-		int maxValue = 20;
+		int maxValue = 30;
 		int testTime = 1000000;
 		System.out.println("测试开始");
 		for (int i = 0; i < testTime; i++) {
