+++ src/class21/Code04_MinCoinsOnePaper.java
@@ -1,4 +1,4 @@
-package class20;
+package class21;
 
 import java.util.HashMap;
 import java.util.Map.Entry;
