+++ src/class22/Code01_BobDie.java
@@ -1,4 +1,4 @@
-package class21;
+package class22;
 
 public class Code01_BobDie {
