+++ src/class22/Code03_SplitSumClosed.java
@@ -1,4 +1,4 @@
-package class21;
+package class22;
 
 import java.util.TreeSet;
