+++ src/class22/Code05_NQueens.java
@@ -1,4 +1,4 @@
-package class21;
+package class22;
 
 public class Code05_NQueens {
